FIX: (#80) 무한 페이징 을 통한 모임 조회 시 N + 1 문제의 개선을 진행한다

+++ src/main/java/com/meetup/server/user/dto/response/UserEventHistoryResponse.java
@@ -16,7 +16,7 @@ public record UserEventHistoryResponse(
         UUID eventId,
         String middlePointName,
         String placeName,
-        int participatedPeopleCount,
+        long participatedPeopleCount,
         List<String> userProfileImageUrls,
         int eventMadeAgo,
         String eventTimeAgo,
